tests/wpt/web-platform-tests/css/css-box/parsing/margin-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing margin with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-margin">
<meta name="assert" content="margin supports only the grammar '[ <length> | <percentage> | auto]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("margin", "available");
test_invalid_value("margin", "10px border-box");
test_invalid_value("margin", "1% 2% 3% 4% 5%");

test_invalid_value("margin-top", "calc(2em + 3ex) auto");
test_invalid_value("margin-right", "auto calc(2em + 3ex) 20%");
test_invalid_value("margin-bottom", "10px 20% calc(2em + 3ex) auto");
test_invalid_value("margin-bottom-left", "none");
</script>
</body>
</html>
